third_party/blink/web_tests/platform/generic/storage/quota/storagequota-request-quota-expected.txt

This tests requesting a quota using Quota API.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS grantedQuota is 0
PASS successfullyParsed is true

TEST COMPLETE
